HWKAPM-517 Move 'trace' specific MDBs and JMS publishers to a sub-package, in preparation for adding similar processing capabilities for Spans
