+++ server/jms/src/main/java/org/hawkular/apm/server/jms/trace/FragmentCompletionTimeDeriverMDB.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.apm.server.jms;
+package org.hawkular.apm.server.jms.trace;
 
 import javax.annotation.PostConstruct;
 import javax.ejb.ActivationConfigProperty;
@@ -23,47 +23,49 @@ import javax.inject.Inject;
 import javax.jms.MessageListener;
 
 import org.hawkular.apm.api.model.events.CompletionTime;
-import org.hawkular.apm.processor.tracecompletiontime.TraceCompletionInformation;
-import org.hawkular.apm.processor.tracecompletiontime.TraceCompletionTimeDeriver;
+import org.hawkular.apm.api.model.trace.Trace;
+import org.hawkular.apm.processor.fragmentcompletiontime.FragmentCompletionTimeDeriver;
+import org.hawkular.apm.server.jms.FragmentCompletionTimePublisherJMS;
+import org.hawkular.apm.server.jms.RetryCapableMDB;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 
 /**
  * @author gbrown
  */
-@MessageDriven(name = "TraceCompletionInformation_TraceCompletionTimeDeriver",
+@MessageDriven(name = "Trace_FragmentCompletionTimeDeriver",
         messageListenerInterface = MessageListener.class,
         activationConfig =
         {
                 @ActivationConfigProperty(propertyName = "destinationType", propertyValue = "javax.jms.Topic"),
-                @ActivationConfigProperty(propertyName = "destination", propertyValue = "TraceCompletionInformation"),
+                @ActivationConfigProperty(propertyName = "destination", propertyValue = "Traces"),
                 @ActivationConfigProperty(propertyName = "subscriptionDurability", propertyValue = "Durable"),
-                @ActivationConfigProperty(propertyName = "clientID", propertyValue = TraceCompletionTimeDeriverMDB.SUBSCRIBER),
+                @ActivationConfigProperty(propertyName = "clientID", propertyValue = FragmentCompletionTimeDeriverMDB.SUBSCRIBER),
                 @ActivationConfigProperty(propertyName = "subscriptionName",
-                            propertyValue = TraceCompletionTimeDeriverMDB.SUBSCRIBER),
-                @ActivationConfigProperty(propertyName = "messageSelector", propertyValue = "subscriber IS NULL OR subscriber = '"+TraceCompletionTimeDeriverMDB.SUBSCRIBER+"'")
+                            propertyValue = FragmentCompletionTimeDeriverMDB.SUBSCRIBER),
+                @ActivationConfigProperty(propertyName = "messageSelector", propertyValue = "subscriber IS NULL OR subscriber = '"+FragmentCompletionTimeDeriverMDB.SUBSCRIBER+"'")
         })
-public class TraceCompletionTimeDeriverMDB extends RetryCapableMDB<TraceCompletionInformation, CompletionTime> {
+public class FragmentCompletionTimeDeriverMDB extends RetryCapableMDB<Trace, CompletionTime> {
 
     @Inject
-    private TraceCompletionInformationPublisherJMS traceCompletionInformationPublisher;
+    private TracePublisherJMS tracePublisher;
 
     @Inject
-    private TraceCompletionTimePublisherJMS traceCompletionTimePublisher;
+    private FragmentCompletionTimePublisherJMS fragmentCompletionTimePublisher;
 
     /**  */
-    public static final String SUBSCRIBER = "TraceCompletionTimeDeriver";
+    public static final String SUBSCRIBER = "FragmentCompletionTimeDeriver";
 
-    public TraceCompletionTimeDeriverMDB() {
+    public FragmentCompletionTimeDeriverMDB() {
         super(SUBSCRIBER);
     }
 
     @PostConstruct
     public void init() {
-        setProcessor(new TraceCompletionTimeDeriver());
-        setRetryPublisher(traceCompletionInformationPublisher);
-        setPublisher(traceCompletionTimePublisher);
-        setTypeReference(new TypeReference<java.util.List<TraceCompletionInformation>>() {
+        setProcessor(new FragmentCompletionTimeDeriver());
+        setRetryPublisher(tracePublisher);
+        setPublisher(fragmentCompletionTimePublisher);
+        setTypeReference(new TypeReference<java.util.List<Trace>>() {
         });
     }
