+++ server/jms/src/main/java/org/hawkular/apm/server/jms/trace/ProducerInfoCacheMDB.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.apm.server.jms;
+package org.hawkular.apm.server.jms.trace;
 
 import java.util.List;
 import java.util.logging.Level;
