+++ server/jms/src/main/java/org/hawkular/apm/server/jms/trace/TraceCompletionInformationInitiatorMDB.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.apm.server.jms;
+package org.hawkular.apm.server.jms.trace;
 
 import javax.annotation.PostConstruct;
 import javax.ejb.ActivationConfigProperty;
@@ -22,46 +22,52 @@ import javax.ejb.MessageDriven;
 import javax.inject.Inject;
 import javax.jms.MessageListener;
 
-import org.hawkular.apm.api.model.events.Notification;
 import org.hawkular.apm.api.model.trace.Trace;
-import org.hawkular.apm.processor.notification.NotificationDeriver;
+import org.hawkular.apm.processor.tracecompletiontime.TraceCompletionInformation;
+import org.hawkular.apm.processor.tracecompletiontime.TraceCompletionInformationInitiator;
+import org.hawkular.apm.server.jms.RetryCapableMDB;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 
 /**
  * @author gbrown
  */
-@MessageDriven(name = "Trace_NotificationDeriver", messageListenerInterface = MessageListener.class,
+@MessageDriven(name = "Trace_TraceCompletionInformationInitiator",
+        messageListenerInterface = MessageListener.class,
         activationConfig =
         {
                 @ActivationConfigProperty(propertyName = "destinationType", propertyValue = "javax.jms.Topic"),
                 @ActivationConfigProperty(propertyName = "destination", propertyValue = "Traces"),
                 @ActivationConfigProperty(propertyName = "subscriptionDurability", propertyValue = "Durable"),
-                @ActivationConfigProperty(propertyName = "clientID", propertyValue = NotificationDeriverMDB.SUBSCRIBER),
-                @ActivationConfigProperty(propertyName = "subscriptionName", propertyValue = NotificationDeriverMDB.SUBSCRIBER),
-                @ActivationConfigProperty(propertyName = "messageSelector", propertyValue = "subscriber IS NULL OR subscriber = '"+NotificationDeriverMDB.SUBSCRIBER+"'")
+                @ActivationConfigProperty(propertyName = "clientID",
+                            propertyValue = TraceCompletionInformationInitiatorMDB.SUBSCRIBER),
+                @ActivationConfigProperty(propertyName = "subscriptionName",
+                            propertyValue = TraceCompletionInformationInitiatorMDB.SUBSCRIBER),
+                @ActivationConfigProperty(propertyName = "messageSelector", propertyValue = "subscriber IS NULL OR subscriber = '"+TraceCompletionInformationInitiatorMDB.SUBSCRIBER+"'")
         })
-public class NotificationDeriverMDB extends RetryCapableMDB<Trace, Notification> {
+public class TraceCompletionInformationInitiatorMDB
+                extends RetryCapableMDB<Trace, TraceCompletionInformation> {
 
     @Inject
     private TracePublisherJMS tracePublisher;
 
     @Inject
-    private NotificationPublisherJMS notificationPublisher;
+    private TraceCompletionInformationPublisherJMS traceCompletionInformationPublisher;
 
     /**  */
-    public static final String SUBSCRIBER = "NotificationDeriver";
+    public static final String SUBSCRIBER = "TraceCompletionInformationInitiator";
 
-    public NotificationDeriverMDB() {
+    public TraceCompletionInformationInitiatorMDB() {
         super(SUBSCRIBER);
     }
 
     @PostConstruct
     public void init() {
-        setProcessor(new NotificationDeriver());
+        setProcessor(new TraceCompletionInformationInitiator());
         setRetryPublisher(tracePublisher);
-        setPublisher(notificationPublisher);
+        setPublisher(traceCompletionInformationPublisher);
         setTypeReference(new TypeReference<java.util.List<Trace>>() {
         });
     }
+
 }
