+++ server/jms/src/main/java/org/hawkular/apm/server/jms/trace/TraceCompletionInformationPublisherJMS.java
@@ -14,10 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.apm.server.jms;
+package org.hawkular.apm.server.jms.trace;
 
 import org.hawkular.apm.processor.tracecompletiontime.TraceCompletionInformation;
 import org.hawkular.apm.processor.tracecompletiontime.TraceCompletionInformationPublisher;
+import org.hawkular.apm.server.jms.AbstractPublisherJMS;
 
 /**
  * This class represents the trace completion information JMS publisher.
