+++ server/jms/src/main/resources/META-INF/services/org.hawkular.apm.api.services.TracePublisher
@@ -1,1 +1,1 @@
-org.hawkular.apm.server.jms.TracePublisherJMS
+org.hawkular.apm.server.jms.trace.TracePublisherJMS
